collect_users/methods/papers_with_code/requirements.txt
pandas==1.1.5
beautifulsoup4==4.10.0
paperswithcode-client==0.3.1